backend/requirements.txt
fastapi==0.105.0
uvicorn[standard]==0.25.0
pydantic==2.5.0
pydantic-settings==2.1.0
qdrant-client==1.8.0
openai==1.3.5
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic-core==2.14.5
aiofiles==23.2.1
pytest==7.4.3
httpx==0.25.2
python-multipart==0.0.6